Shape the rift around a four-pointed star
The tear now opens into a star at its centre, tracking the seam so it stays attached as the seam wavers, and its tails run past the top and bottom edges instead of terminating on screen.

The star is defined once in star.glsl, prepended to the fragment shaders that need it, and a RIFT_STAR_ONLY define reduces the rift to the bare star for previewing the shape. The WebGL helpers move to gl.ts so more than one canvas can share them.
